Consolidate panels
