--- lib/widgets/common/panels/developer_panel.dart
@@ -1,12 +1,0 @@
-import 'package:bratacha/intl/localizations.dart';
-import 'package:bratacha/widgets/common/text/clickable_text.dart';
-import 'package:flutter/material.dart';
-
-class DeveloperPanel extends StatelessWidget {
-  const DeveloperPanel({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return ClickableText(context.l10n.developerPanelLabel);
-  }
-}
