+++ test/widgets/home_screen/settings_tab/credits_panel_test.dart
@@ -1,5 +1,5 @@
 import 'package:bratacha/intl/localizations.dart';
-import 'package:bratacha/widgets/common/panels/developer_panel.dart';
+import 'package:bratacha/widgets/common/text/clickable_text.dart';
 import 'package:bratacha/widgets/home_screen/settings_tab/credits_panel.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -16,7 +16,7 @@ void main() {
       expect(find.byType(Column), findsOneWidget);
       expect(find.byType(Text), findsOneWidget);
       expect(find.byType(SizedBox), findsOneWidget);
-      expect(find.byType(DeveloperPanel), findsOneWidget);
+      expect(find.byType(ClickableText), findsOneWidget);
     });
 
     testWidgets('Ensure content is correct', (tester) async {
